django-api/requirements.txt
asgiref==3.6.0
Django==3.2.16
django-crispy-forms==1.14.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
requests
mysqlclient
django-crontab
